+++ src/test/java/org/stellar/sdk/ServerTest.java
@@ -30,49 +30,6 @@ import org.stellar.sdk.responses.TransactionResponse;
 import org.stellar.sdk.responses.operations.OperationResponse;
 
 public class ServerTest {
-  private final String publicRootResponse =
-      "{\n"
-          + "  \"_links\": {\n"
-          + "    \"account\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/accounts/{account_id}\",\n"
-          + "      \"templated\": true\n"
-          + "    },\n"
-          + "    \"account_transactions\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/accounts/{account_id}/transactions{?cursor,limit,order}\",\n"
-          + "      \"templated\": true\n"
-          + "    },\n"
-          + "    \"assets\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/assets{?asset_code,asset_issuer,cursor,limit,order}\",\n"
-          + "      \"templated\": true\n"
-          + "    },\n"
-          + "    \"metrics\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/metrics\"\n"
-          + "    },\n"
-          + "    \"order_book\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/order_book{?selling_asset_type,selling_asset_code,selling_asset_issuer,buying_asset_type,buying_asset_code,buying_asset_issuer,limit}\",\n"
-          + "      \"templated\": true\n"
-          + "    },\n"
-          + "    \"self\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/\"\n"
-          + "    },\n"
-          + "    \"transaction\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/transactions/{hash}\",\n"
-          + "      \"templated\": true\n"
-          + "    },\n"
-          + "    \"transactions\": {\n"
-          + "      \"href\": \"https://horizon.stellar.org/transactions{?cursor,limit,order}\",\n"
-          + "      \"templated\": true\n"
-          + "    }\n"
-          + "  },\n"
-          + "  \"horizon_version\": \"0.18.0-92259749c681df66a8347f846e94681a24f2a920\",\n"
-          + "  \"core_version\": \"stellar-core 11.1.0 (324c1bd61b0e9bada63e0d696d799421b00a7950)\",\n"
-          + "  \"history_latest_ledger\": 24345129,\n"
-          + "  \"history_elder_ledger\": 1,\n"
-          + "  \"core_latest_ledger\": 24345130,\n"
-          + "  \"network_passphrase\": \"Public Global Stellar Network ; September 2015\",\n"
-          + "  \"current_protocol_version\": 11,\n"
-          + "  \"core_supported_protocol_version\": 11\n"
-          + "}";
   private final String successResponse =
       "{\n"
           + "  \"_links\": {\n"
@@ -1351,6 +1308,8 @@ public class ServerTest {
       assertEquals(400, e.getCode().intValue());
     }
   }
+
+  public void testLoadAccount() {}
 
   private Dispatcher buildTestCheckMemoRequiredMockDispatcher() {
     final String memoRequiredResponse =