+++ src/test/java/org/stellar/sdk/operations/ManageBuyOfferOperationTest.java
@@ -4,7 +4,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.stellar.sdk.Asset.create;
 
-import java.io.IOException;
 import java.math.BigDecimal;
 import org.junit.Test;
 import org.stellar.sdk.Asset;
@@ -12,11 +11,10 @@ import org.stellar.sdk.AssetTypeCreditAlphaNum4;
 import org.stellar.sdk.AssetTypeNative;
 import org.stellar.sdk.KeyPair;
 import org.stellar.sdk.Price;
-import org.stellar.sdk.exception.StrKeyException;
 
 public class ManageBuyOfferOperationTest {
   @Test
-  public void testManageBuyOfferOperation() throws IOException, StrKeyException {
+  public void testManageBuyOfferOperation() {
     // GC5SIC4E3V56VOHJ3OZAX5SJDTWY52JYI2AFK6PUGSXFVRJQYQXXZBZF
     KeyPair source =
         KeyPair.fromSecretSeed("SC4CGETADVYTCR5HEAVZRB3DZQY5Y4J7RFNJTRA6ESMHIPEZUSTE2QDK");
@@ -61,7 +59,7 @@ public class ManageBuyOfferOperationTest {
   }
 
   @Test
-  public void testManageBuyOfferOperationWithConstructorPrice() throws StrKeyException {
+  public void testManageBuyOfferOperationWithConstructorPrice() {
     // See https://github.com/stellar/java-stellar-sdk/issues/292
     // GC5SIC4E3V56VOHJ3OZAX5SJDTWY52JYI2AFK6PUGSXFVRJQYQXXZBZF
     KeyPair source =